Added formal regression test script run_regress, which creates files regress_diffs and regress_log.  Email to Rob and Brett as follows:
I have written a formal regression test script for starcheck.  It is now
part of the starcheck CVS distribution.  The script is a /bin/sh script
called 'run_regress'.  This script runs the test version of starcheck (in
the current directory) on a number of loads, specified in the script, and
runs the release starcheck on those same loads.  The starcheck outputs
(starcheck.txt) are diffed and saved in a file called 'regress_diffs'.
The processing output is saved in 'regress_log'.

Standard test procedure is now something like:

- Check out starcheck from CVS
- Edit starcheck, do preliminary testing to verify expected functionality
- Possibly edit run_regress to add a load which tests the changes
- 'make test'
- Inspect regress_diffs and verify expected output
- Possibly inspect HTML starcheck outputs within the 'regress' directory
- Present to load review for approval
- 'cvs update'
- 'cvs commit'  (put in meaningful change comments here)
- 'make install'

Note that the files regress_diffs and regress_log are also part of the
CVS tree, so the regression outputs for each revision are automatically
captured.