Changes from 4.27 ( missing log messages )
See 2006-AprStarcheckRelease on the Wiki.

*  Improvements
   o Added new information to Obsid summary of output including:
       + Count of warnings per obsid
       + Count of warning-free ACQ and GUI stars
   o Reformatted warnings to begin with Index and then specify type and details
   o Continued conversion to IO::All for file access
   o Implemented new AGASC confirmation checks:
       + Confirm that there is an AGASC star near assigned search box centers
for BOT/GUI/ACQ
       + Confirm that the assigned AGASC star exists at the proper Y-ang, Z-ang

* Bug Fixes

   o Fixed observation duration bug: Starcheck was using the maneuver duration
as the observation duration when checking dither. Now the observation interval
between the end of the last maneuver and the beginning of the next is used (For
the final observation of the load, the end of the science observation is used as
the end of the observation).
   o Fixed catalog area bug. Stars were missing at the corners of the starcheck
plots. Increased the size of the AGASC cone to fix this.