In preparation for 2006-Sep release, parse_cm_file.pl was removed and replaced with this package.
Changes:
--------

Now contains new routine verifyMM() that creates an array of the maneuver
commands from the backstop and compares them to the maneuvers in the maneuver
summary.

MM() has been totally revised to allow it to return either an array or hash of
maneuvers. Parsing of the maneuver summary is also now completed in "paragraphs"
instead of line by line. The suffix of intermediate attitudes (in "segmented"
maneuvers) has been changed to "_IA" from "!" .