Updated dark cal to do per-replica checking.  Reorganized manvr/dwell checking.  Added momentum unload checking.
The dark cal checker was configured to match:

a) a single template command list (for one transponder) with the planned list
b) a list of maneuver pointings with planned pointings
c) a list of maneuver dwells with planned calculated dwells

The first check was broken for transponder switches (which are not
infrequent).  It was also broken for the split-over-perigee case if
any commands occurred during the gap that could be mistaken for dark
current commanding (CIMODESL)

The matches for b) and c) have not been a problem, but would have
failed for a dark current with pointings away from DFC during a split.

Thus the hex command checking has been broken into a check for each
replica.  The transponder and dither state are also checked before
each replica.

The manvr and dwell checks have been relaxed to ignore any extra
dwells at DFC (which should occur if there are pointings away from DFC
during a split).  However, it is now explicitly checked that maneuvers
to and from a replica are always from and to DFC.

Additionally, a new sub has been added to check for momentum unloads
during dark cal operations.