Reduced DOT vs Backstop command time slop allowance from ~3min to 1sec.
Starcheck was still configured to work on old products with a 3 minute
DOT vs Backstop offset.  This was a problem for the APR0411A loads, as
it contained an obsid with a 4 second duration, and the matching in
get_obsid() managed to assign the Backstop MP_OBSID commanding into
the wrong Obsid object.  Reduced allowed DOT/Backstop slop to fix.
Updated regression tests to remove loads before the 3 minute offset
was removed.