Include regression test changes due to star DB lookups
Setting starcheck to use the database for all warning message results
in many changes in the starcheck output.  This is expected.  In
previous regression testing, the output file warnings on stars were
based on the content of the bad acqs rdb, which is up-to-date through
2008.

For observations before the last update to that table, the data in
that table does not help to recreate the data that would have been
available when these catalogs were evaluated.  For example the DEC2704
old output for obsid 5425 references 2/2 bad acquisitions for star
166461744.  Neither of those acquisitions had occurred at the time of
observation (one of them actually *is* obsid 5425).

The database lookup filters the guide and acquisition histories based
on the observation date, so only observations before the checked
observation change its warnings.

Additionally, as the database is perpetually up-to-date through last
week, observations since 2008 now show more accurate text warnings.

Differences have been spot checked and look good.