Fix bug in execution order to make set_npm_times work
Previously, the components of set_npm_times() were in the check_dither
and check_momentum subroutines and were called from the main checking
loop in starcheck.pl.  Attempts to call set_npm_times the main setup
routine (the loop that has set_maneuver() etc) failed mostly silently
because neither the {next} pointer nor the actual maneuver data for
the next obsid had been populated.  This commit puts the call to
set_npm_times just before it is used by the dither and momentum
checks, and adds a comment to the subroutine.