Write pgplot png files directly
Do not make "pgplot.png" and rename.  While pgplot.png is probably
created on the same filesystem as the intended plot, that is not
guaranteed, so 'rename' may be unsafe.  Also, it is perhaps unwise to
have the three plotting routines sharing the same-named temporary
file; this change eliminates the use of the temporary file altogether.