Add a checks to confirm non-zero-size plots
Let each plotting routine check that it has successfully created a
plot file and that file has non-zero-size.  When pgplot fails to open
or write a file, it writes to STDERR but does not die or
return good status info (pgbegin seems to always return 1).  Without
explicit file checks, there could be a failure to create the plots
without a corresponding warning() in STDERR and in the processing
warnings in the starcheck output.  This changes adds those checks.