Fix OR parsing of TARGET_OFFSET
The target offset regular expression was returning results like:

TARGET_OFFSET_Z = "ROLL=(127.0"

as the regular expression was not restrictive enough for the case when
the target offset is specified as a single value and the ROLL data follows
on the same line.  I believe that the single value case is just the
TARGET_OFFSET_Y value (judging from OCAT y_det_offset lookups), so I've
broken the single value case out to its own regex.