Put out per-obsid temperatures from aca_check
Instead of a time-list of temperatures over the interval, use the NPNT
intervals in the commanded states to map each obsid to a max temp
over the interval in python and read those temperatures from Perl.
The Perl still needs to do the "does the next obsid have a maneuver"
check.