Fix uninitialized errors on 2014/JAN2514/oflsa
Because the observation tstop is not known/set due to the replan/reopen,
the check_bright_perigee routine was throwing a

Use of uninitialized value in numeric gt (>) at src/lib/Ska/Starcheck/Obsid.pm line 719.

for every line in the radmon history.  This change just warns and exits
if the observation has no obs_tstop.