Reduce benign dark cal transponder warnings
The code was set up to warn on any transponder commanding that changed
the apparent transponder state during a replica.  This included, say,
CIMODESL commands that were consistent with the selected replica, but
not consistent with the load-commanded state before the replica.  The new
code tracks "state" changes as well as the states and only throws warnings
on new commands if they are actually inconsistent with the transponder that
is expected for the replica.