Set main checking to use configurable min obsid defining ERs
Some of the Obsid.pm code was set to handle obsids > 40000 as ERs,
some was set > 50000.  This change sets ER_MIN to 38000 and uses that
through most of the code.  $self->{is_science} would have been a good
candidate too, but this change wasn't worth the code reorg to make sure
that would be defined for each sub that needed this.