Reorganize dither checking
The dither checking code was exiting early for ERs and not doing some
checks that can be done for ERs.  It was also exciting for ERs before
setting values for dither parameters when the are known.  These changes
add checks on ERs for dither change during observation and for large dither.
These changes also set dither parameters in the Obsid object for ERs if
they are known for an ER so that ERs won't all be checked using 20 arcsec
dither (if there is dither commanding in Backstop before the ER being checked).