Move dither setting a bit later in check_star_catalog
The new warnings are popping up on Dark Cal replicas.  While
a better plan would be to just not call check_star_catalog if there
is no star catalog, we can get rid of the dither warnings by just
hitting the check for MP_STARCAT before setting and checking that
there is reasonable dither assigned to the observation.